Another CurrencyFrame bugfix
Changed the code that stops iterating when index is larger than  GetCurrencyListSize to follow the previous pattern of error checking before executing instead of returning.  Returning early prevents the code that collapses headers that were expanded from running.  This fix should no longer cause all currency headers to be remain expanded when the iterator is called.